Some basic bracket matching dataset creation/training
Issues: This dataset counts the number of unmatched brackets. This
means that the “classifications” actually aren’t classifications and
they’re not one-hot encoded. This makes things like the confusion
matrix and our current measures of accuracy a little pointless. It
would probably make more sense to evaluate the efficacy based solely on
the quadratic cost?